# Array Encoding/Decoding performance improvements