--- a/src/Encoding/IntEncoding.cs
+++ b/src/Encoding/IntEncoding.cs
@@ -3,7 +3,12 @@
 
 namespace Microsoft.Azure.Amqp.Encoding
 {
-    sealed class IntEncoding : EncodingBase
+    using System;
+    using System.Buffers.Binary;
+    using System.Collections;
+    using System.Collections.Generic;
+
+    sealed class IntEncoding : PrimitiveEncoding
     {
         public IntEncoding()
             : base(FormatCode.Int)
@@ -85,6 +90,56 @@ namespace Microsoft.Azure.Amqp.Encoding
         public override object DecodeObject(ByteBuffer buffer, FormatCode formatCode)
         {
             return IntEncoding.Decode(buffer, formatCode);
+        }
+
+        public override int GetArrayEncodeSize(IList value)
+        {
+            return FixedWidth.Int * value.Count;
+        }
+
+        public override void EncodeArray(IList value, ByteBuffer buffer)
+        {
+            int byteCount = FixedWidth.Int * value.Count;
+
+            buffer.Validate(write: true, byteCount);
+            Span<byte> destination = buffer.GetWriteSpan();
+            
+            if (value is int[] intArray)
+            {
+                // fast-path for int[] so the bounds checks can be elided
+                for (int i = 0; i < intArray.Length; i++)
+                {
+                    BinaryPrimitives.WriteInt32BigEndian(destination.Slice(FixedWidth.Int * i), intArray[i]);
+                }
+            }
+            else
+            {
+                IReadOnlyList<int> listValue = (IReadOnlyList<int>)value;
+                for (int i = 0; i < listValue.Count; i++)
+                {
+                    BinaryPrimitives.WriteInt32BigEndian(destination.Slice(FixedWidth.Int * i), listValue[i]);
+                }
+            }
+
+            buffer.Append(byteCount);
+        }
+
+        public override Array DecodeArray(ByteBuffer buffer, int count, FormatCode formatCode)
+        {
+            int byteCount = FixedWidth.Int * count;
+
+            buffer.Validate(write: false, byteCount);
+            ReadOnlySpan<byte> source = buffer.GetReadSpan();
+            
+            int[] array = new int[count];
+            for (int i = 0; i < count; ++i)
+            {
+                array[i] = BinaryPrimitives.ReadInt32BigEndian(source.Slice(FixedWidth.Int * i));
+            }
+
+            buffer.Complete(byteCount);
+
+            return array;
         }
     }
 }