@xinchen10 2020-11-13
I like the idea and the results looks good. The current code tries to reuse the other encodings for array items but has to pay the price of boxing/unboxing for primitive types. Looking at the int32 encoding I guess we will need to create other primitive encoding types to handle array encoding. If we have to have type specific code, maybe we can keep all of them in one place, e.g. the ArrayEncoding class. Let me look at this area a bit more.
The same issue exists in list encoding as well, I believe. So if we have an approach that also works for list, it would be great.