@xinchen10 2020-11-14
@eerhardt Hi, the approach looks good to me. We can continue with other primitive types. The list encoding only matters when the items are of the same type so we don't have to worry about that for now. Thanks.